requirements.txt
#general requirements
discord.py
python-dotenv
asyncio
#music bot requirements
spotipy
PyNaCl
ffmpeg

#gpt-3 chatbot requirements
openai
requests